reorganize code, tests
