+++ stac_api/common/models/links.py
