+++ stac_api/elastic_search/config.py
