+++ stac_api/elastic_search/session.py
