+++ stac_api/elastic_search/setup.cfg
