+++ stac_api/elastic_search/setup.py
@@ -65,6 +65,6 @@ setup(
     tests_require=extra_reqs["dev"],
     extras_require=extra_reqs,
     entry_points={
-        "console_scripts": ["stac-fastapi-elasticsearch=elastic_search.app:run"]
+        "console_scripts": ["stac-fastapi-elasticsearch=stac_api.elastic_search.app:run"]
     },
 )
