+++ stac_api/open_search/setup.cfg
