+++ stac_api/tests/extensions/cql2/example04.json
