+++ stac_api/tests/extensions/cql2/example05a.json
