+++ stac_api/tests/extensions/cql2/example06b.json
