+++ stac_api/tests/extensions/cql2/example08.json
