+++ stac_api/tests/extensions/cql2/example10.json
