+++ stac_api/tests/extensions/cql2/example18.json
