+++ stac_api/tests/extensions/cql2/example19.json
