+++ stac_api/tests/extensions/cql2/example20.json
