+++ stac_api/tests/extensions/cql2/example22.json
