+++ stac_api/tests/resources/test_collection.py
